feat: reboot and poweroff actions for linux agents
