+++ internal/agent/agent.go
@@ -14,7 +14,6 @@ import (
 	"os/user"
 	"path/filepath"
 	"slices"
-	"strconv"
 	"strings"
 	"time"
 
@@ -588,69 +587,6 @@ func (a *Agent) StopVNCSubscribe() error {
 
 	if err != nil {
 		return fmt.Errorf("[ERROR]: could not subscribe to agent stop vnc, reason: %v", err)
-	}
-	return nil
-}
-
-func (a *Agent) RebootSubscribe() error {
-	_, err := a.NATSConnection.QueueSubscribe("agent.reboot."+a.Config.UUID, "openuem-agent-management", func(msg *nats.Msg) {
-		log.Println("[INFO]: reboot request received")
-		if err := msg.Respond([]byte("Reboot!")); err != nil {
-			log.Printf("[ERROR]: could not respond to agent reboot message, reason: %v\n", err)
-		}
-
-		action := openuem_nats.RebootOrRestart{}
-		if err := json.Unmarshal(msg.Data, &action); err != nil {
-			log.Printf("[ERROR]: could not unmarshal to agent reboot message, reason: %v\n", err)
-			return
-		}
-
-		when := int(time.Until(time.Now()).Seconds())
-		if when > 0 {
-			if err := exec.Command("cmd", "/C", "shutdown", "/r", "/t", strconv.Itoa(when)).Run(); err != nil {
-				fmt.Printf("[ERROR]: could not initiate power off, reason: %v", err)
-			}
-		} else {
-			if err := exec.Command("cmd", "/C", "shutdown", "/r").Run(); err != nil {
-				fmt.Printf("[ERROR]: could not initiate power off, reason: %v", err)
-			}
-		}
-	})
-
-	if err != nil {
-		return fmt.Errorf("[ERROR]: could not subscribe to agent reboot, reason: %v", err)
-	}
-	return nil
-}
-
-func (a *Agent) PowerOffSubscribe() error {
-	_, err := a.NATSConnection.QueueSubscribe("agent.poweroff."+a.Config.UUID, "openuem-agent-management", func(msg *nats.Msg) {
-		log.Println("[INFO]: power off request received")
-		if err := msg.Respond([]byte("Power Off!")); err != nil {
-			log.Printf("[ERROR]: could not respond to agent power off message, reason: %v\n", err)
-			return
-		}
-
-		action := openuem_nats.RebootOrRestart{}
-		if err := json.Unmarshal(msg.Data, &action); err != nil {
-			log.Printf("[ERROR]: could not unmarshal to agent power off message, reason: %v\n", err)
-			return
-		}
-
-		when := int(time.Until(time.Now()).Seconds())
-		if when > 0 {
-			if err := exec.Command("cmd", "/C", "shutdown", "/s", "/t", strconv.Itoa(when)).Run(); err != nil {
-				fmt.Printf("[ERROR]: could not initiate power off, reason: %v", err)
-			}
-		} else {
-			if err := exec.Command("cmd", "/C", "shutdown", "/s").Run(); err != nil {
-				fmt.Printf("[ERROR]: could not initiate shutdown, reason: %v", err)
-			}
-		}
-	})
-
-	if err != nil {
-		return fmt.Errorf("[ERROR]: could not subscribe to agent power off, reason: %v", err)
 	}
 	return nil
 }
