+++ internal/agent/agent_windows.go
@@ -6,6 +6,9 @@ import (
 	"encoding/json"
 	"fmt"
 	"log"
+	"os/exec"
+	"strconv"
+	"time"
 
 	"github.com/nats-io/nats.go"
 	openuem_nats "github.com/open-uem/nats"
@@ -59,6 +62,69 @@ func (a *Agent) StartVNCSubscribe() error {
 
 	if err != nil {
 		return fmt.Errorf("[ERROR]: could not subscribe to agent start vnc, reason: %v", err)
+	}
+	return nil
+}
+
+func (a *Agent) RebootSubscribe() error {
+	_, err := a.NATSConnection.QueueSubscribe("agent.reboot."+a.Config.UUID, "openuem-agent-management", func(msg *nats.Msg) {
+		log.Println("[INFO]: reboot request received")
+		if err := msg.Respond([]byte("Reboot!")); err != nil {
+			log.Printf("[ERROR]: could not respond to agent reboot message, reason: %v\n", err)
+		}
+
+		action := openuem_nats.RebootOrRestart{}
+		if err := json.Unmarshal(msg.Data, &action); err != nil {
+			log.Printf("[ERROR]: could not unmarshal to agent reboot message, reason: %v\n", err)
+			return
+		}
+
+		when := int(time.Until(action.Date).Seconds())
+		if when > 0 {
+			if err := exec.Command("cmd", "/C", "shutdown", "/r", "/t", strconv.Itoa(when)).Run(); err != nil {
+				fmt.Printf("[ERROR]: could not initiate power off, reason: %v", err)
+			}
+		} else {
+			if err := exec.Command("cmd", "/C", "shutdown", "/r").Run(); err != nil {
+				fmt.Printf("[ERROR]: could not initiate power off, reason: %v", err)
+			}
+		}
+	})
+
+	if err != nil {
+		return fmt.Errorf("[ERROR]: could not subscribe to agent reboot, reason: %v", err)
+	}
+	return nil
+}
+
+func (a *Agent) PowerOffSubscribe() error {
+	_, err := a.NATSConnection.QueueSubscribe("agent.poweroff."+a.Config.UUID, "openuem-agent-management", func(msg *nats.Msg) {
+		log.Println("[INFO]: power off request received")
+		if err := msg.Respond([]byte("Power Off!")); err != nil {
+			log.Printf("[ERROR]: could not respond to agent power off message, reason: %v\n", err)
+			return
+		}
+
+		action := openuem_nats.RebootOrRestart{}
+		if err := json.Unmarshal(msg.Data, &action); err != nil {
+			log.Printf("[ERROR]: could not unmarshal to agent power off message, reason: %v\n", err)
+			return
+		}
+
+		when := int(time.Until(action.Date).Seconds())
+		if when > 0 {
+			if err := exec.Command("cmd", "/C", "shutdown", "/s", "/t", strconv.Itoa(when)).Run(); err != nil {
+				fmt.Printf("[ERROR]: could not initiate power off, reason: %v", err)
+			}
+		} else {
+			if err := exec.Command("cmd", "/C", "shutdown", "/s").Run(); err != nil {
+				fmt.Printf("[ERROR]: could not initiate shutdown, reason: %v", err)
+			}
+		}
+	})
+
+	if err != nil {
+		return fmt.Errorf("[ERROR]: could not subscribe to agent power off, reason: %v", err)
 	}
 	return nil
 }
