+++ internal/service/linux/main.go
@@ -4,6 +4,9 @@ package main
 
 import (
 	"log"
+	"os"
+	"os/signal"
+	"syscall"
 
 	"github.com/open-uem/openuem-agent/internal/commands/report"
 	"github.com/open-uem/openuem-agent/internal/logger"
@@ -23,4 +26,9 @@ func main() {
 		log.Println("error running report")
 	}
 	r.Print()
+
+	// TODO LINUX Keep the connection alive
+	done := make(chan os.Signal, 1)
+	signal.Notify(done, syscall.SIGINT, syscall.SIGTERM, os.Interrupt)
+	<-done
 }
